Inicialización ventana de ogre
