+++ Src/MotorGrafico/GraphicsEngine.cpp
@@ -1,5 +1,9 @@
 #include "GraphicsEngine.h"
 #include <OgreRoot.h>
+#include <SDL.h>
+#include <SDL_video.h>
+#include <SDL_syswm.h>
+
 
 GraphicsEngine* GraphicsEngine::instance = nullptr;
 
@@ -19,20 +23,13 @@ GraphicsEngine* GraphicsEngine::getInstance()
 	return instance;
 }
 
-void GraphicsEngine::test() {
-#ifdef _DEBUG
-	root = new Ogre::Root("OgreDEBUG/pluginsDEBUG.cfg");
-#else //RELEASE
-	root = new Ogre::Root("OgreRELEASE/pluginsRELEASE.cfg");
-#endif
-}
 
 void GraphicsEngine::initRoot()
 {
 #ifdef _DEBUG
-	root = new Ogre::Root("OgreDEBUG/pluginsDEBUG.cfg");
+	root = new Ogre::Root("OgreDEBUG/pluginsDEBUG.cfg", "OgreDEBUG/ogreDEBUG.cfg");
 #else //RELEASE
-	root = new Ogre::Root("OgreRELEASE/pluginsRELEASE.cfg");
+	root = new Ogre::Root("OgreRELEASE/pluginsRELEASE.cfg", "OgreRELEASE/ogreRELEASE.cfg");
 #endif
 }
 
@@ -41,8 +38,27 @@ void GraphicsEngine::initWindow() {
 	root->initialise(false);
 	Ogre::NameValuePairList params;
 	Ogre::ConfigOptionMap configuracion = root->getRenderSystem()->getConfigOptions();
+	
+	SDL_Init(SDL_INIT_EVERYTHING);
+	
+	SDL_SysWMinfo wmInfo;
+	SDL_VERSION(&wmInfo.version);
+	
+	Uint32 flags = SDL_WINDOW_ALLOW_HIGHDPI; //SDL_WINDOW_RESIZABLE
+	std::string nombre = "Prueba";
 
-	window = root->createRenderWindow("Prueba", 1920, 1080, false, &params);
+	sdlWindow = SDL_CreateWindow(nombre.c_str(), SDL_WINDOWPOS_UNDEFINED, SDL_WINDOWPOS_UNDEFINED, 1280, 720, flags); 
+	SDL_GetWindowWMInfo(sdlWindow, &wmInfo);
+	
+	params["FSAA"] = configuracion["FSAA"].currentValue;     
+	params["vsync"] = configuracion["VSync"].currentValue;
+	params["gamma"] = configuracion["sRGB Gamma Conversion"].currentValue;
+	params["externalWindowHandle"] = Ogre::StringConverter::toString(size_t(wmInfo.info.win.window));
 
+	window = root->createRenderWindow("PruebaOgre", 1920, 1080, false, &params);
+}
+
+void GraphicsEngine::start()
+{
 	root->startRendering();
 }
