+++ bin/OgreDEBUG/ogreDEBUG.cfg
