+++ bin/OgreRELEASE/ogreRELEASE.cfg
@@ -1,0 +1,29 @@
+Render System=Direct3D11 Rendering Subsystem
+
+[Direct3D11 Rendering Subsystem]
+Allow NVPerfHUD=No
+Backbuffer Count=Auto
+Driver type=Hardware
+FSAA=1
+Floating-point mode=Fastest
+Full Screen=No
+Information Queue Exceptions Bottom Level=Info (exception on any message)
+Max Requested Feature Levels=11.0
+Min Requested Feature Levels=9.1
+Rendering Device=(default)
+Reversed Z-Buffer=No
+VSync=Yes
+VSync Interval=1
+Video Mode=800 x 600 @ 32-bit colour
+sRGB Gamma Conversion=No
+
+[OpenGL Rendering Subsystem]
+Colour Depth=
+Display Frequency=N/A
+FSAA=0
+Full Screen=No
+RTT Preferred Mode=FBO
+VSync=Yes
+VSync Interval=1
+Video Mode= 640 x  480
+sRGB Gamma Conversion=No
